änderungen in functionPoint berechnung (casts)

+++ src/model/services/calculation/FunctionPoint.java
@@ -84,7 +84,7 @@ public class FunctionPoint extends Calculation{
 				}
 			}
 		}
-		return unweigtedFP;
+		return (double)unweigtedFP;
 	}
 
 	/**
